refactor: Rename Patient Session Route
