+++ src/features/patient-session/patient-session.controller.ts
@@ -1,23 +1,56 @@
-import { Handler, Request } from 'express';
+import { Handler, Router } from 'express';
 
-import PatientSessionService, { Options } from './patient-session.service';
+import * as PatientService from './patient-session.service';
 
-export const parseRequest = (req: Request): Options => ({
-    count: parseInt(<string>req.query.count || '500'),
-    page: parseInt(<string>req.query.page || '0'),
-    start: <string>req.query.start,
-    end: <string>req.query.end,
-    patientName: req.query.patientName
-        ? decodeURI(<string>req.query.patientName)
-        : undefined,
+const patientSessionController = Router();
+const patientSessionSummaryController = Router();
+
+const getController =
+    (service: PatientService.Service): Handler =>
+    (req, res) => {
+        service(req.snowflake, req.options)
+            .then((data) => res.json({ data }))
+            .catch((err) => res.status(500).json({ error: err.message }));
+    };
+
+patientSessionController.use((req, res, next) => {
+    req.options = {
+        count: parseInt(<string>req.query.count || '500'),
+        page: parseInt(<string>req.query.page || '0'),
+        start: <string>req.query.start,
+        end: <string>req.query.end,
+        patientName: req.query.patientName
+            ? decodeURI(<string>req.query.patientName)
+            : undefined,
+    };
+
+    next();
 });
 
-const PatientSessionController: Handler = async (req, res) => {
-    const options = parseRequest(req);
+patientSessionController.get('/', getController(PatientService.getAll));
 
-    PatientSessionService(req.snowflake, options)
-        .then((data) => res.json({ data }))
-        .catch((err) => res.status(500).json({ error: err.message }));
-};
+patientSessionSummaryController.get('/', getController(PatientService.getCount));
 
-export default PatientSessionController;
+patientSessionSummaryController.get(
+    '/start-of-month',
+    getController(PatientService.getCountByStartOfMonth),
+);
+
+patientSessionSummaryController.get(
+    '/compliant',
+    getController(PatientService.getCountByCompliant),
+);
+
+patientSessionSummaryController.get(
+    '/therapy-mode-group',
+    getController(PatientService.getCountByTherapyModeGroup),
+);
+
+patientSessionSummaryController.get(
+    '/age',
+    getController(PatientService.getCountByAge),
+);
+
+patientSessionController.use('/summary', patientSessionSummaryController);
+
+export default patientSessionController;
