+++ src/features/patient-session/patient-session.repository.test.ts
@@ -1,0 +1,10 @@
+import PatientSessionRepository from './patient-session.repository';
+import cases from './patient-session.config.test';
+
+describe('Build', () => {
+    it.each(cases)('$name', async ({ options }) => {
+        const query = PatientSessionRepository(options).select().toQuery();
+        console.log({ query });
+        expect(query).toBeTruthy();
+    });
+});
