+++ src/index.ts
@@ -2,8 +2,8 @@ import { http } from '@google-cloud/functions-framework';
 import express from 'express';
 
 import { getConnection } from './providers/snowflake';
-import PatientSessionController from './features/patient-session/patient-session.controller';
-import patientController from './features/patient/patient.controller';
+
+import patientSessionController from './features/patient-session/patient-session.controller';
 
 const app = express();
 
@@ -25,7 +25,6 @@ app.use((req, res, next) => {
     }
 });
 
-app.get('/patient-session', PatientSessionController);
-app.use('/patient', patientController);
+app.get('/patient-session', patientSessionController);
 
 http('main', app);
