+++ src/types/express.d.ts
@@ -1,7 +1,7 @@
 import { Request } from 'express';
 import { Connection } from 'snowflake-sdk';
 
-import { Options } from '../features/patient/patient.service';
+import { Options } from '../features/pati../features/patient-session/patient.service
 
 declare module 'express-serve-static-core' {
     interface Request {
